Move code from StyledCell to an Extension file
