+++ Classes/Views/StyledTableCell.swift
@@ -19,15 +19,6 @@ class StyledTableCell: UITableViewCell {
         super.init(style: style, reuseIdentifier: reuseIdentifier)
         configure()
     }
-    
-    func setup(with file: RepositoryFile) {
-        textLabel?.text = file.name
-        
-        let imageName = file.isDirectory ? "file-directory" : "file"
-        imageView?.image = UIImage(named: imageName)?.withRenderingMode(.alwaysTemplate)
-        imageView?.tintColor = Styles.Colors.blueGray.color
-        accessoryType = file.isDirectory ? .disclosureIndicator : .none
-    }
 
     private func configure() {
         textLabel?.font = Styles.Fonts.body
